backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-jose[cryptography]
httpx
youtube-transcript-api
crewai
langchain-openai
